nd: Fix identification of data across platforms
When hashing the navigation data on e.g. WASM and iOS, we get different
results and thus can't reliably identify the navigation data
partitions. Therefore, hash the data used to create the navigation data
and store the partition ID. This retains the same partition ID when
serializing and deserializing the navigation data on different
platforms.

Signed-off-by: Joe Pearson <joe@aeronauticalmaps.com>