fms: Add evaluation pipeline
The pipeline provides a clear order in which the FMS needs to be update
after one of its sub-systems changed. The pipeline can skip stages that
were not affected by a change, but always runs the stages in the same
order.

This should avoid errors where the FMS is not re-evaluated in the right
order or stages are missed when adding new sub-systems.

Signed-off-by: Joe Pearson <joe@aeronauticalmaps.com>