route: Change runway prompt to fix ambiguity
Attaching the runway to the airport word simplifies parsing
significantly and we always have a direct mapping of runway and airport
where we can validate the runway already during lexing.

Furthermore, this fixes a premature assignment of landing runway during
token parsing in case an intermediate airport was defined with a runway
but the destination not. Checking those prompt errors is now
significantly easier with an airport/runway word.

Signed-off-by: Joe Pearson <joe@aeronauticalmaps.com>